refactor: move some mock extensions to the manager
