+++ JotunnLib/Managers/MockManager.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 using UnityEngine.SceneManagement;
 using Object = UnityEngine.Object;
 
@@ -7,7 +8,7 @@ namespace Jotunn.Managers
     /// <summary>
     ///     Handles all logic to do with managing mocked prefabs added into the game.
     /// </summary>
-    class MockManager : IManager
+    internal class MockManager : IManager
     {
         private static MockManager _instance;
         /// <summary>
@@ -21,9 +22,26 @@ namespace Jotunn.Managers
                 return _instance;
             }
         }
+        
+        /// <summary>
+        ///     Legacy ValheimLib prefix used by the Mock System to recognize Mock gameObject that must be replaced at some point.
+        /// </summary>
+        [Obsolete("Legacy ValheimLib mock prefix. Use JVLMockPrefix \"JVLmock_\" instead.")]
+        public const string MockPrefix = "VLmock_";
 
+        /// <summary>
+        ///     Prefix used by the Mock System to recognize Mock gameObject that must be replaced at some point.
+        /// </summary>
+        public const string JVLMockPrefix = "JVLmock_";
+
+        /// <summary>
+        ///     Internal container for mocked prefabs
+        /// </summary>
         internal GameObject MockPrefabContainer;
 
+        /// <summary>
+        ///     Creates the container and registers all hooks
+        /// </summary>
         public void Init()
         {
             MockPrefabContainer = new GameObject("MockPrefabs");
@@ -33,9 +51,9 @@ namespace Jotunn.Managers
             On.ObjectDB.Awake += RemoveMockPrefabs;
         }
 
-        internal T CreateMockedPrefab<T>(string prefabName) where T : Component
+        public T CreateMockedPrefab<T>(string prefabName) where T : Component
         {
-            string name = PrefabExtension.JVLMockPrefix + prefabName;
+            string name = JVLMockPrefix + prefabName;
 
             Transform transform = MockPrefabContainer.transform.Find(name);
             if (transform != null)
@@ -77,5 +95,60 @@ namespace Jotunn.Managers
                 }
             }
         }
+
+        /// <summary>
+        ///     Will try to find the real vanilla prefab from the given mock
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="unityObject"></param>
+        /// <returns>the real prefab</returns>
+        public static T GetRealPrefabFromMock<T>(Object unityObject) where T : Object
+        {
+            return (T)GetRealPrefabFromMock(unityObject, typeof(T));
+        }
+        
+#pragma warning disable CS0618
+        /// <summary>
+        ///     Will try to find the real vanilla prefab from the given mock
+        /// </summary>
+        /// <param name="unityObject"></param>
+        /// <param name="mockObjectType"></param>
+        /// <returns>the real prefab</returns>
+        public static Object GetRealPrefabFromMock(Object unityObject, Type mockObjectType)
+        {
+            if (unityObject)
+            {
+                var unityObjectName = unityObject.name;
+                var isVLMock = unityObjectName.StartsWith(MockPrefix);
+                var isJVLMock = unityObjectName.StartsWith(JVLMockPrefix);
+                if (isVLMock || isJVLMock)
+                {
+                    if (isVLMock) unityObjectName = unityObjectName.Substring(MockPrefix.Length);
+                    if (isJVLMock) unityObjectName = unityObjectName.Substring(JVLMockPrefix.Length);
+
+                    // Cut off the suffix in the name to correctly query the original material
+                    if (unityObject is Material)
+                    {
+                        const string materialInstance = " (Instance)";
+                        if (unityObjectName.EndsWith(materialInstance))
+                        {
+                            unityObjectName = unityObjectName.Substring(0, unityObjectName.Length - materialInstance.Length);
+                        }
+                    }
+
+                    Object ret = PrefabManager.Cache.GetPrefab(mockObjectType, unityObjectName);
+
+                    if (!ret)
+                    {
+                        throw new Exception($"Mock prefab {unityObjectName} could not be resolved");
+                    }
+
+                    return ret;
+                }
+            }
+
+            return null;
+        }
+#pragma warning restore CS0618
     }
 }
